requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
google-generativeai
python-dotenv